Caching error caused me to try a bunch of stuff I didn't need to
