--- index.ios.js
@@ -1,6 +1,0 @@
-
-var PlayInstallReferrer = {};
-
-PlayInstallReferrer.getInstallReferrerInfo = function() {};
-
-module.exports = { PlayInstallReferrer }
